requirements.txt
flask==3.1.2
flask-cors==3.0.10
gunicorn==23.0.0
openai>=1.0.0
numpy==1.26.4
faiss-cpu>=1.7.4; platform_system != "Windows"
python-dotenv==1.2.1
requests==2.32.5
tqdm==4.66.1
